refactor(engine): move transfer_test.go's tests beside the code they exercise
transfer_test.go had become a catch-all: 22 tests for six source files, so a
reader looking for the sanitizer tests had to skip past the protocol matrix
and the stats report to find them. Each test now lives in the test file named
for the source it exercises: safeJoin, sanitizeComponent and the displayText
drift guard in sanitize_test.go; parseMetadata, classifyControl and
looksLikeJSONObject in control_test.go; Summarize, formatBytes, truncateName
and displayText in format_test.go; CheckCompat and the compat messages
appended to protocol_test.go; TestClaimPart beside its de-collision siblings
in decollide_test.go; TestCommitPart beside the staged-write tests in
partial_test.go. What remains, TestReportBytesToServer, is renamed
receiver_test.go to pair with reportBytesToServer's file. The new files carry
no GOOS token, and their notes sit below the package clause.

Pure move: the sorted go test -list output is identical on Windows (81 names)
and GOOS=linux (78 names, the cross-compiled binary run under WSL), the -v
PASS list is identical (217 names), go doc -all -u is unchanged, and git diff
--color-moved shows only package, import and header lines as new.
